+++ bundle/src/projects/commercial/modules/article-body-adverts.ts
@@ -1,7 +1,8 @@
 import type { AdSize, SizeMapping } from '@guardian/commercial-core';
 import { adSizes, createAdSlot } from '@guardian/commercial-core';
 import { isInUk } from 'common/modules/commercial/geo-utils';
-import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid';
+import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid-check';
+
 import { createAdvertBorder } from 'common/modules/spacefinder-debug-tools';
 import {
 	getCurrentBreakpoint,
