+++ bundle/src/projects/commercial/modules/liveblog-adverts.ts
@@ -1,7 +1,7 @@
 import { adSizes, createAdSlot } from '@guardian/commercial-core';
 import { log } from '@guardian/libs';
 import { isInUk } from 'common/modules/commercial/geo-utils';
-import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid';
+import { isInEagerPrebidVariant } from 'common/modules/experiments/tests/eager-prebid-check';
 import { getCurrentBreakpoint } from 'lib/detect-breakpoint';
 import { getUrlVars } from 'lib/url';
 import fastdom from '../../../lib/fastdom-promise';
