+++ bundle/src/projects/common/modules/experiments/tests/eager-prebid.ts
@@ -1,7 +1,4 @@
 import type { ABTest } from '@guardian/ab-core';
-import { memoize } from 'lodash-es';
-import { isInUk } from 'common/modules/commercial/geo-utils';
-import { getSynchronousTestsToRun } from '../ab';
 import { bypassMetricsSampling } from '../utils';
 
 export const eagerPrebid: ABTest = {
@@ -25,14 +22,3 @@ export const eagerPrebid: ABTest = {
 	],
 	canRun: () => true,
 };
-
-// determine if the user is in any of the the eager prebid variants
-export const isInEagerPrebidVariant = memoize((): boolean => {
-	if (!isInUk()) {
-		return false;
-	}
-	const test = getSynchronousTestsToRun().find(
-		(test) => test.id === eagerPrebid.id,
-	);
-	return test ? test.variantToRun.id !== 'control' : false;
-});
